chore: optimize dev loading time and support indexer-less app
