+++ tailwind.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: ["./index.html", "./src/**/*.{ts,tsx,js,jsx}"],
+};
